backend/requirements.txt
# ============================================
# AI Doctor Platform - Backend Dependencies
# ============================================
# Web Framework
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ============================================
# Data Validation & Settings
# ============================================
pydantic>=2.8.0
pydantic-settings>=2.4.0

# ============================================
# Environment
# ============================================
python-dotenv==1.0.1

# ============================================
# Supabase (Database & Storage)
# ============================================
supabase==2.5.0

# ============================================
# AI/ML - Google Gemini
# ============================================
google-generativeai==0.7.0

# ============================================
# OCR - Text Extraction from PDFs/Images
# ============================================
pytesseract==0.3.13
pdfplumber==0.11.4

# ============================================
# Image Processing
# ============================================
Pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0

# ============================================
# Pose Estimation (for posture analysis)
# ============================================
mediapipe>=0.10.30

# ============================================
# PDF Generation
# ============================================
reportlab==4.2.2

# ============================================
# File Handling (Async)
# ============================================
aiofiles==24.1.0

# ============================================
# HTTP Client
# ============================================
httpx==0.27.0

# ============================================
# Utilities
# ============================================
python-dateutil==2.9.0
setuptools>=65.0.0
